NP-3212: restructure code
