+++ customer-commons/src/main/java/no/unit/nva/customer/Constants.java
@@ -1,11 +1,19 @@
 package no.unit.nva.customer;
 
+import com.amazonaws.auth.DefaultAWSCredentialsProviderChain;
+import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.google.common.net.MediaType;
 import java.net.URI;
 import java.util.List;
+import no.unit.nva.customer.service.CustomerService;
+import no.unit.nva.customer.service.impl.DynamoDBCustomerService;
 import nva.commons.apigateway.MediaTypes;
 import nva.commons.core.Environment;
+import nva.commons.core.JacocoGenerated;
 
+@JacocoGenerated
 public final class Constants {
 
     public static final Environment ENVIRONMENT = new Environment();
@@ -18,10 +26,27 @@ public final class Constants {
     public static final String AWS_REGION = ENVIRONMENT.readEnvOpt("AWS_REGION").orElse(DEFAULT_AWS_REGION);
     public static final String AWD_DYNAMODB_SERVICE_END_POINT = dynamoDbServiceEndpoint();
 
-    private static String dynamoDbServiceEndpoint() {
-        return URI.create(String.format("https://dynamodb.%s.amazonaws.com",AWS_REGION)).toString();
+    private Constants() {
     }
 
-    private Constants() {
+    @JacocoGenerated
+    public static CustomerService defaultCustomerService() {
+        AmazonDynamoDB client = defaultDynamoDbClient();
+        return new DynamoDBCustomerService(client, ObjectMapperConfig.objectMapper, ENVIRONMENT);
+    }
+
+    private static String dynamoDbServiceEndpoint() {
+        return URI.create(String.format("https://dynamodb.%s.amazonaws.com", AWS_REGION)).toString();
+    }
+
+    @JacocoGenerated
+    private static AmazonDynamoDB defaultDynamoDbClient() {
+        EndpointConfiguration endpointConfiguration =
+            new EndpointConfiguration(no.unit.nva.customer.Constants.AWD_DYNAMODB_SERVICE_END_POINT,
+                                      no.unit.nva.customer.Constants.AWS_REGION);
+        return AmazonDynamoDBClientBuilder.standard()
+            .withCredentials(new DefaultAWSCredentialsProviderChain())
+            .withEndpointConfiguration(endpointConfiguration)
+            .build();
     }
 }
