+++ customer-commons/src/main/java/no/unit/nva/customer/model/CustomerDto.java
@@ -2,9 +2,6 @@ package no.unit.nva.customer.model;
 
 import static nva.commons.core.attempt.Try.attempt;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 import java.net.URI;
 import java.time.Instant;
 import java.util.Collections;
@@ -15,12 +12,7 @@ import no.unit.nva.customer.model.interfaces.Context;
 import nva.commons.core.JacocoGenerated;
 import nva.commons.core.JsonUtils;
 
-@SuppressWarnings("PMD.ExcessivePublicCount")
-@JsonTypeInfo(
-    use = JsonTypeInfo.Id.NAME,
-    include = As.PROPERTY,
-    property = "type")
-@JsonTypeName("Customer")
+
 public class CustomerDto extends CustomerDtoWithoutContext implements Context {
 
     @JsonProperty("@context")
