+++ customer-commons/src/main/java/no/unit/nva/customer/model/VocabularySettingDto.java
@@ -1,18 +1,10 @@
 package no.unit.nva.customer.model;
 
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 import java.net.URI;
 import java.util.Objects;
 import no.unit.nva.customer.model.interfaces.VocabularySetting;
 import nva.commons.core.JacocoGenerated;
 
-@JsonTypeInfo(
-    use = JsonTypeInfo.Id.NAME,
-    include = As.PROPERTY,
-    property = "type")
-@JsonTypeName("VocabularySetting")
 public class VocabularySettingDto implements VocabularySetting {
 
     private String name;
