+++ customer-commons/src/main/java/no/unit/nva/customer/model/interfaces/Customer.java
@@ -1,15 +1,17 @@
 package no.unit.nva.customer.model.interfaces;
 
-import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import java.time.Instant;
 import java.util.UUID;
-import no.unit.nva.customer.model.CustomerDao;
-import no.unit.nva.customer.model.CustomerDto;
 
-
+@SuppressWarnings("PMD.ExcessivePublicCount")
+@JsonTypeInfo(
+    use = JsonTypeInfo.Id.NAME,
+    include = As.PROPERTY,
+    property = "type")
+@JsonTypeName("Customer")
 public interface Customer {
 
     UUID getIdentifier();
