+++ customer-commons/src/main/java/no/unit/nva/customer/model/interfaces/VocabularySetting.java
@@ -1,10 +1,16 @@
 package no.unit.nva.customer.model.interfaces;
 
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
 import com.fasterxml.jackson.annotation.JsonTypeName;
+import java.net.URI;
 import no.unit.nva.customer.model.VocabularyStatus;
 
-import java.net.URI;
-
+@SuppressWarnings("PMD.ExcessivePublicCount")
+@JsonTypeInfo(
+    use = JsonTypeInfo.Id.NAME,
+    include = As.PROPERTY,
+    property = "type")
 @JsonTypeName("VocabularySetting")
 public interface VocabularySetting {
 
@@ -19,5 +25,4 @@ public interface VocabularySetting {
     VocabularyStatus getStatus();
 
     void setStatus(VocabularyStatus status);
-
 }
